+++ stroom-app/src/main/java/stroom/startup/App.java
@@ -73,6 +73,7 @@ import stroom.script.server.ScriptServlet;
 import stroom.script.spring.ScriptConfiguration;
 import stroom.search.spring.SearchConfiguration;
 import stroom.security.server.AuthorisationResource;
+import stroom.security.server.JWTService;
 import stroom.security.server.SecurityFilter;
 import stroom.security.server.SessionResource;
 import stroom.security.spring.SecurityConfiguration;
@@ -100,7 +101,7 @@ import stroom.spring.ServerComponentScanConfiguration;
 import stroom.spring.ServerConfiguration;
 import stroom.statistics.server.sql.search.SqlStatisticsQueryResource;
 import stroom.statistics.spring.StatisticsConfiguration;
-import stroom.util.HasHealthCheck;
+import stroom.util.HealthCheckUtils;
 import stroom.util.config.StroomProperties;
 import stroom.util.db.DbUtil;
 import stroom.util.spring.StroomSpringProfiles;
@@ -178,7 +179,7 @@ public class App extends Application<Config> {
         healthCheckRegistry.register(configuration.getProxyConfig().getClass().getName(), new HealthCheck() {
             @Override
             protected Result check() throws Exception {
-                Map<String, Object> detailMap = HasHealthCheck.beanToMap(configuration.getProxyConfig());
+                Map<String, Object> detailMap = HealthCheckUtils.beanToMap(configuration.getProxyConfig());
                 return Result.builder()
                         .healthy()
                         .withDetail("values", detailMap)
@@ -256,6 +257,7 @@ public class App extends Application<Config> {
         SpringUtil.addHealthCheck(environment.healthChecks(), applicationContext, StroomIndexQueryResource.class);
         SpringUtil.addHealthCheck(environment.healthChecks(), applicationContext, DictionaryResource.class);
         SpringUtil.addHealthCheck(environment.healthChecks(), applicationContext, RuleSetResource.class);
+        SpringUtil.addHealthCheck(environment.healthChecks(), applicationContext, JWTService.class);
 
         // Add filters
         SpringUtil.addFilter(servletContextHandler, applicationContext, HttpServletRequestFilter.class, "/*");